Added SSH agent support
The SSH agent is automagically brought up and dropped in login shells.
On logout, if a rogue SSH agent is found, then it is dropped too.

Note that this only works in login shells, and not typical invocations.

Additionally, the setup script has a command to precompile all ZSh
scripts. These precompiled versions are ignored by Git, to prevent any
possible incompatibilities (as well as bloating the repository).